+++ src/aura-historia-api/src/users/types.rs
@@ -1,5 +1,4 @@
 use crate::patch_value::PatchValue;
-use geo::data::address_data::{GeoAddressData, StructuredAddressData};
 use localization::Language;
 use money::Currency;
 use serde::{Deserialize, Serialize};
@@ -41,10 +40,6 @@ pub(crate) struct OwnUserData {
     pub(crate) role: UserRole,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(crate) stripe_customer_id: Option<user_core::stripe_customer_id::StripeCustomerId>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) structured_address: Option<StructuredAddressData>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) geo_address: Option<GeoAddressData>,
 }
 
 impl From<UserDetailsView> for OwnUserData {
@@ -61,8 +56,6 @@ impl From<UserDetailsView> for OwnUserData {
             tier: view.tier,
             role: view.role,
             stripe_customer_id: view.stripe_customer_id,
-            structured_address: view.structured_address.map(Into::into),
-            geo_address: view.geo_address.map(Into::into),
         }
     }
 }
@@ -98,10 +91,6 @@ pub(crate) struct AdminUserData {
     pub(crate) role: UserRole,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub(crate) stripe_customer_id: Option<user_core::stripe_customer_id::StripeCustomerId>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) structured_address: Option<StructuredAddressData>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub(crate) geo_address: Option<GeoAddressData>,
 }
 
 impl From<UserDetailsView> for AdminUserData {
@@ -118,8 +107,6 @@ impl From<UserDetailsView> for AdminUserData {
             tier: view.tier,
             role: view.role,
             stripe_customer_id: view.stripe_customer_id,
-            structured_address: view.structured_address.map(Into::into),
-            geo_address: view.geo_address.map(Into::into),
         }
     }
 }
@@ -189,8 +176,6 @@ pub(crate) struct PatchOwnUserData {
     pub(crate) measurement_unit: PatchValue<MeasurementUnit>,
     #[serde(default)]
     pub(crate) show_unassessed_or_sensitive_content: PatchValue<bool>,
-    #[serde(default)]
-    pub(crate) structured_address: PatchValue<StructuredAddressData>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -229,6 +214,4 @@ pub(crate) struct PatchAdminUserData {
         deserialize_with = "crate::wire::user_role::patch::deserialize"
     )]
     pub(crate) role: PatchValue<UserRole>,
-    #[serde(default)]
-    pub(crate) structured_address: PatchValue<StructuredAddressData>,
 }