+++ src/geo/src/data/mod.rs
@@ -1,2 +1,1 @@
-pub mod address_data;
 pub mod continent_data;