+++ src/product-listing-opensearch/src/product_listing_document.rs
@@ -474,14 +474,6 @@ mod tests {
             "shopName",
             "sellerName",
             "shopType",
-            "structuredAddressAddressline",
-            "structuredAddressAddresslineExtra",
-            "structuredAddressLocality",
-            "structuredAddressRegion",
-            "structuredAddressPostalCode",
-            "structuredAddressCountry",
-            "structuredAddressContinent",
-            "geoAddress",
             "viewUrl",
         ] {
             assert!(