+++ src/product-listing-opensearch/src/product_listing_percolation_document.rs
@@ -663,14 +663,6 @@ mod tests {
             "shopName",
             "sellerName",
             "shopType",
-            "structuredAddressAddressline",
-            "structuredAddressAddresslineExtra",
-            "structuredAddressLocality",
-            "structuredAddressRegion",
-            "structuredAddressPostalCode",
-            "structuredAddressCountry",
-            "structuredAddressContinent",
-            "geoAddress",
         ] {
             assert!(
                 document.get(field).is_none(),